MetricsService/requirement.txt
pytz 
pymongo
matplotlib reportlab
PyGithub
Flask